+++ src/features/usage-consolidation-accordion/usage-consolidation-content-title.js
@@ -113,6 +113,7 @@ const UsageConsolidationContentTitle = ({
 
       return (
         <List
+          marginBottom0
           items={coverageDates}
           listClass={styles.coverageDatesList}
         />
@@ -148,24 +149,18 @@ const UsageConsolidationContentTitle = ({
     }) : [];
 
   const customProperties = {
-    visibleColumns: ['packageName', 'coverage', 'cost', 'usage', 'costPerUse'],
+    visibleColumns: ['packageName', 'coverage', 'cost'],
     columnMapping: {
       packageName: intl.formatMessage({ id: 'ui-eholdings.usageConsolidation.summary.packageName' }),
       coverage: intl.formatMessage({ id: 'ui-eholdings.usageConsolidation.summary.coverage' }),
-      usage: (props) => (
-        <UsageColumnHeader
-          {...props}
-          label={intl.formatMessage({ id: 'ui-eholdings.usageConsolidation.summary.title.usage' })}
-        />
-      ),
+      usage: null,
+      costPerUse: null,
       cost: 'ui-eholdings.usageConsolidation.summary.titleCost',
     },
     columnWidths: {
-      packageName: '33%',
-      coverage: '33%',
-      cost: '12%',
-      usage: '10%',
-      costPerUse: '12%',
+      packageName: '43%',
+      coverage: '39%',
+      cost: '18%',
     },
     formatter,
   };