+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 * Support `eholdings` interface version `3.0`.
 * Support erm interface v4.0
 * Fix typos in Full text requests table title. (UIEH-1039)
+* Title record: Usage & analysis > Summary Table > Remove Usage and CostPerUse columns. (UIEH-1040)
 
 ## [5.0.0] (https://github.com/folio-org/ui-eholdings/tree/v5.0.0) (2020-10-15)